Split components into CommonJS modules
