+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules/
-public/js-prod/
+public/js/
 config.json
 npm-debug.log
