+++ package.json
@@ -18,7 +18,8 @@
     "express": "^4.12.3",
     "node-tweet-stream": "^1.8.1",
     "react": "^0.13.3",
-    "socket.io": "^1.3.5"
+    "socket.io": "^1.3.5",
+    "socket.io-client": "^1.3.5"
   },
   "license": "MIT",
   "devDependencies": {
