+++ public/index.html
@@ -1,5 +1,5 @@
 <!doctype html>
-<html class="no-js" lang="">
+<html class="no-js" lang="en">
 <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -17,8 +17,6 @@
     <h1 class="container slogan">Twitter Stream</h1>
     <div class="container" id="TwitterStreamWrapper"></div>
 
-    <script src="https://cdn.socket.io/socket.io-1.2.0.js"></script>
-    <script src="http://fb.me/react-0.12.1.js"></script>
-    <script type="text/javascript" src="js-prod/twitter-stream-app.js"></script>
+    <script src="js/app.min.js"></script>
 </body>
 </html>
