+++ src/app.js
@@ -1,0 +1,9 @@
+'use strict';
+
+var React = require('react');
+var TwitterStreamApp = require('./components/stream-app');
+
+React.render(
+    <TwitterStreamApp />,
+    document.getElementById('TwitterStreamWrapper')
+);
